Refactor code & fix minor bugs
