+++ src/client-desktop/CMakeLists.txt
@@ -48,6 +48,7 @@ set(SOURCES
     src/chat_list_model.cpp
     src/chat_list_view.cpp
     src/chat_widget.cpp
+    src/db_storage.cpp
     src/main.cpp
     src/main_window.cpp
     src/message_item_delegate.cpp
