+++ src/client-desktop/include/chat_widget.hpp
@@ -66,6 +66,7 @@ private:  // NOLINT (readability-redundant-access-specifiers)
     chat_handle_t m_current_chat_it;
     QSet<int> m_pressed_keys;
     QString m_incomplete_message;
+    QString m_pre_edit_message;
     int m_current_chat_row;
     int m_edit_row;
     bool m_edit_mode = false;
