+++ src/client-desktop/include/ram_storage.hpp
@@ -11,8 +11,6 @@
 #include <utility>
 #include <vector>
 
-#include <iostream>
-
 namespace melon::client_desktop
 {
