+++ src/client-desktop/src/main.cpp
@@ -28,12 +28,19 @@ QString create_connection_with_db()
     {
         QSqlQuery query(db);
 
+        if (!query.exec(QStringLiteral("PRAGMA foreign_keys = ON")))
+        {
+            std::cout << "Error: " << query.lastError().text().toStdString() << std::endl;
+            return error(qApp->translate("CreateDB", "Couldn't create messages table: %1"));
+        }
+
         if(!query.exec(QStringLiteral("CREATE TABLE [chats]"
                                       "(chat_id int NOT NULL,"
                                       " domain_id INT NOT NULL,"
                                       " name TEXT,"
                                       " PRIMARY KEY (chat_id, domain_id))")))
         {
+            std::cout << "Error: " << query.lastError().text().toStdString() << std::endl;
             return error(qApp->translate("CreateDB", "Couldn't create chats table"));
         }
 
@@ -46,7 +53,7 @@ QString create_connection_with_db()
                                       " text TEXT,"
                                       " status INT, "
                                       " PRIMARY KEY (message_id, chat_id, domain_id),"
-                                      " FOREIGN KEY (chat_id) REFERENCES [chats] (chat_id) ON DELETE CASCADE)")))
+                                      " FOREIGN KEY (chat_id, domain_id) REFERENCES chats(chat_id, domain_id) ON DELETE CASCADE)")))
         {
             std::cout << "Error: " << query.lastError().text().toStdString() << std::endl;
             return error(qApp->translate("CreateDB", "Couldn't create messages table: %1"));
